web-sentimenter/pyproject.toml
[tool.poetry]
name = "web-sentimenter"
version = "0.1.0"
description = ""
authors = ["trisnawahyudiii <wayantrisna79@gmail.com>"]
readme = "README.md"
packages = [{include = "main"}]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
flask-cors = "^4.0.0"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.0"
flask-bcrypt = "^1.0.1"
pymysql = "^1.1.0"
cryptography = "^41.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
